+++ src/main/java/ar/edu/unc/david/petrinetsimulator/log/PetriLogPatterns.java
@@ -1,4 +1,4 @@
-package ar.edu.unc.david.petrinetsimulator;
+package ar.edu.unc.david.petrinetsimulator.log;
 
 import java.util.regex.Pattern;
