+++ src/main/java/ar/edu/unc/david/petrinetsimulator/monitor/MonitorInterface.java
@@ -1,4 +1,4 @@
-package ar.edu.unc.david.petrinetsimulator;
+package ar.edu.unc.david.petrinetsimulator.monitor;
 
 /** Interface for monitoring the firing of transitions in a Petri net. */
 public interface MonitorInterface {
