+++ src/main/java/ar/edu/unc/david/petrinetsimulator/policy/Policy.java
@@ -1,4 +1,4 @@
-package ar.edu.unc.david.petrinetsimulator;
+package ar.edu.unc.david.petrinetsimulator.policy;
 
 /** Interface for defining policies to select which transition to fire among candidates. */
 public interface Policy {
